app/requirements.txt
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
autopep8==1.7.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.79.0
flower==1.2.0
h11==0.13.0
httptools==0.4.0
humanize==4.3.0
idna==3.3
kombu==5.2.4
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycodestyle==2.9.1
pydantic==1.9.2
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tornado==6.2
typing-extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
vine==5.0.0
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
